+++ test/python_helpers.py
@@ -7,3 +7,10 @@ def delete_arg(arg):
 def single_dict_arg(d):
   if not isinstance(d, dict):
     raise Exception('argument is not a dict')
+
+def catch_exception(callable):
+  try:
+    callable()
+    return None
+  except Exception as err:
+    return err